requirements.txt
absl-py==2.1.0
astunparse==1.6.3
blinker==1.7.0
certifi==2024.2.2
chardet==5.2.0
charset-normalizer==2.0.12
click==8.1.7
cmake==3.29.3
colorama==0.4.6
dnspython==2.6.1
easyocr==1.7.1
filelock==3.13.3
Flask==2.0.3
flatbuffers==24.3.25
fsspec==2024.3.1
gast==0.5.4
google-pasta==0.2.0
grpcio==1.62.1
gunicorn==20.1.0
h5py==3.10.0
huggingface-hub==0.23.4
idna==3.6
imageio==2.34.0
itsdangerous==2.1.2
Jinja2==3.1.3
keras==3.1.1
lazy_loader==0.4
libclang==18.1.1
Markdown==3.6
markdown-it-py==3.0.0
MarkupSafe==2.1.5
mdurl==0.1.2
ml-dtypes==0.3.2
mpmath==1.3.0
namex==0.0.7
networkx==3.3
ninja==1.11.1.1
numpy==1.26.4
opencv-python==4.9.0.80
opencv-python-headless==4.9.0.80
opt-einsum==3.3.0
optree==0.11.0
packaging==24.0
pillow==10.2.0
protobuf==4.25.3
pyclipper==1.3.0.post5
Pygments==2.17.2
pymongo==4.6.3
pytesseract==0.3.10
python-bidi==0.4.2
PyYAML==6.0.1
pyzbar==0.1.9
regex==2024.5.15
reportlab==4.1.0
requests==2.26.0
rich==13.7.1
safetensors==0.4.3
scikit-image==0.22.0
scipy==1.13.0
setuptools==69.2.0
shapely==2.0.3
six==1.16.0
sympy==1.12
tensorboard==2.16.2
tensorboard-data-server==0.7.2
tensorflow==2.16.1
tensorflow-intel==2.16.1
termcolor==2.4.0
tifffile==2024.2.12
tokenizers==0.19.1
torch==2.2.2
torchvision==0.17.2
tqdm==4.66.4
transformers==4.41.2
typing_extensions==4.10.0
urllib3==1.26.19
Werkzeug==2.0.3
wheel==0.43.0
wrapt==1.16.0
